generalize job count analytics
