+++ AppBookDataAnalyzer/Analytics for Rows/Career Explorer/JobExploreCount.swift
@@ -8,53 +8,12 @@
 import Foundation
 
 struct JobExploreCount: Analytic {
-    
-    let exploredJobEventCode = 39
-    
+        
     var title: String = "Job - Explored Count"
     
     func analyze(database: Database, textbookMaterial: TextbookMaterial) async -> String? {
         
-        try? await database.pool.read { db in
-            
-            switch textbookMaterial {
-                
-            case .page(let appbook, let pageNumber):
-                
-                let query = """
-                    SELECT COUNT(*)
-                    FROM \(Database.EventLog.tableName)
-                    WHERE \(Database.EventLog.Column.appbookId) = \(appbook.id)
-                    AND \(Database.EventLog.Column.pageNumber) = \(pageNumber)
-                    AND \(Database.EventLog.Column.code) = \(exploredJobEventCode)
-                """
-                
-                guard let count = try Int.fetchOne(db, sql: query) else {
-                    return nil
-                }
-                
-                return String(count)
-                
-                
-            case .job(let job):
-                
-                let query = """
-                    SELECT COUNT(*)
-                    FROM \(Database.EventLog.tableName)
-                    WHERE \(Database.EventLog.Column.code) = \(exploredJobEventCode)
-                    AND \(Database.EventLog.Column.contextId) = \(job.id)
-                """
-                
-                guard let count = try Int.fetchOne(db, sql: query) else {
-                    return nil
-                }
-                
-                return String(count)
-                
-            }
-            
-            
-        }
+        await Database.Job.analyze(eventCode: 39, database: database, textbookMaterial: textbookMaterial)
         
     }
