+++ AppBookDataAnalyzer/Analytics for Rows/Career Explorer/JobMainImageViews.swift
@@ -8,52 +8,12 @@
 import Foundation
 
 struct JobMainImageViews: Analytic {
-    
-    let jobMainImageEventCode = 40
-    
+        
     var title: String = "Job - Main Image View Count"
     
     func analyze(database: Database, textbookMaterial: TextbookMaterial) async -> String? {
         
-        try? await database.pool.read { db in
-            
-            switch textbookMaterial {
-                
-            case .page(let appbook, let pageNumber):
-                
-                let query = """
-                    SELECT COUNT(*)
-                    FROM \(Database.EventLog.tableName)
-                    WHERE \(Database.EventLog.Column.appbookId) = \(appbook.id)
-                    AND \(Database.EventLog.Column.pageNumber) = \(pageNumber)
-                    AND \(Database.EventLog.Column.code) = \(jobMainImageEventCode)
-                """
-                
-                guard let count = try Int.fetchOne(db, sql: query) else {
-                    return nil
-                }
-                
-                return String(count)
-                
-                
-            case .job(let job):
-                
-                let query = """
-                    SELECT COUNT(*)
-                    FROM \(Database.EventLog.tableName)
-                    WHERE \(Database.EventLog.Column.code) = \(jobMainImageEventCode)
-                    AND \(Database.EventLog.Column.contextId) = \(job.id)
-                """
-                
-                guard let count = try Int.fetchOne(db, sql: query) else {
-                    return nil
-                }
-                
-                return String(count)
-                
-            }
-            
-        }
+        await Database.Job.analyze(eventCode: 40, database: database, textbookMaterial: textbookMaterial)
         
     }
